\name{baseball}
\alias{baseball}
\docType{data}
\title{
Baseball data.
}
\description{
Age and weight data for 1015 major league baseball players.
}
\usage{data(baseball)}
\format{
  A data frame with 1015 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{Age.}
    \item{\code{y}}{Weight.}
  }
}
\references{
Matloff,  N. (2017) "Statistical  regression  and  classification:  From  linear  models  to  machinelearning". Chapman and Hall/CRC.
}
\keyword{datasets}
