\name{bone}
\alias{bone}
\docType{data}
\title{
Bone mineral density data.
}
\description{
The data set contains measurements on the relative change in (spinal) bone mineral density over one year for 485 North
American adolescents.
}
\usage{data(bone)}
\format{
  A data frame with 485 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{Age of the subject.}
    \item{\code{y}}{Relative change in (spinal) bone mineral density.}
  }
}
\references{
Bachrach et al., (1999) "Bone mineral acquisition in healthy Asian, Hispanic, black, and Caucasian youth: a longitudinal study".
The journal of clinical endocrinology & metabolism.
}
\keyword{datasets}
