\name{boxOffice}
\alias{boxOffice}
\docType{data}
\title{
Movie Box-office Revenue Data
}
\description{
Box-office revenues during opening and after the first week.
}
\usage{data(boxOffice)}
\format{
  A data frame with 4031 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{Log of opening box-office revenues.}
    \item{\code{y}}{Log of box-office revenues after the first week.}
  }
}
\references{
Voudouris et al., (2012) "Modelling skewness and kurtosis with the BCPE density in GAMLSS". Journal of Applied Statistics, 39(6), 1279-1293.
}
\keyword{datasets}
