\name{bupa}
\alias{bupa}
\docType{data}
\title{
BUPA liver disorders data.
}
\description{
A modified version of BUPA liver disorders data set, containing measurements of gamma-glutamyl transpeptidase (GGT) and alanine-aminotransferase (ALT)
extracted from 345 male individuals' blood sample.
}
\usage{data(bupa)}
\format{
  A data frame with 345 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{Log of gamma-glutamyl transpeptidase.}
    \item{\code{y}}{Log of alanine-aminotransferase.}
  }
}
\references{
McDermott, J. & Forsyth, R.S. (2016) "Diagnosing a disorder in a classification benchmark". Pattern Recognition Letters, 73, 41-43.
}
\keyword{datasets}
