\name{butterfly}
\alias{butterfly}
\docType{data}
\title{
The Butterfly data.
}
\description{
The stylized simulated example used in our paper. 
}
\usage{data(butterfly)}
\format{
  A data frame with 700 observations on 2 variables.
  \describe{
    \item{\code{x}}{Values of covariate \eqn{X}.}
    \item{\code{y}}{Values of \eqn{Y}.}
  }
}
\references{
Mukhopadhyay, S., and Wang, K (2020) "Breiman's `Two Cultures' Revisited and Reconciled". Technical Report. 
}
\keyword{datasets}
