\name{cholesterol}
\alias{cholesterol}
\docType{data}
\title{
LDL cholesterol of Quail.
}
\description{
Completely randomized experiment investigating LDL (low-density lipoprotein) cholesterol in
quails.
}
\usage{data(cholesterol)}
\format{
  A data frame with 39 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{Type of diet, each is mixed with a different drug compound.}
    \item{\code{y}}{Measurments of LDL cholesterol levels}
  }
}
\references{
Hettmansperger, T. P. and J. W. McKean (2010), "Robust nonparametric statistical methods", CRC Press.
}
\keyword{datasets}
