\name{dutch}
\alias{dutch}
\docType{data}
\title{
Dutch Boys BMI data
}
\description{
This dataset is a part of the Fourth Dutch Growth Study, which comprised of observations on age and BMI of 7294 Dutch boys. A slightly modified version that is available in \code{gamlss.data} package.
}
\usage{data(dutch)}
\format{
  A data frame with 7294 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{Subject age.}
    \item{\code{y}}{Subject BMI.}
  }
}
\references{
Fredriks et al., (2000)
"Body index measurements in 1996-7 compared with 1980". Archives of disease in childhood 82(2), 107-112.
}
\keyword{datasets}
