\name{LPMachineLearning-package}
\alias{LPMachineLearning-package}
\alias{LPMachineLearning}
\docType{package}
\title{
\packageTitle{LPMachineLearning}
}
\description{
This package provides a unified interface to convert any black-box ML regression algorithms into an exploratory 
uncertainty prediction machine that is robust, interpretable, and scalable for large datasets. A large variety 
of modeling and predictive inference tasks can be done using the fitted model.
}

\author{
\packageAuthor{LPMachineLearning}

Maintainer: \packageMaintainer{LPMachineLearning}
}
\references{
Mukhopadhyay, S., and Wang, K (2020) "Breiman's 'Two Cultures' Revisited and Reconciled". Technical Report. 
}
\keyword{ package }

